Block proposing moved from Node to Consensus.
